refactor(UserInput.js): make class to file
